requirements.txt
SpeechRecognition
# google-cloud-speech
# google-cloud-storage
# google-cloud-core
groq
dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
dateparser
# openai-whisper
# torch
# torchaudio
# --index-url https://download.pytorch.org/whl/cu118

